Create README.md
This repository contains a VBA script that reads through the data from an excel file with yearly stock exchange information. It is able to read through a sorted list of stock names and calculates the yearly price change, the percent change, and the total volume of stocks for each name. It also highlights a positive price change with a green color and a negative with a red.